requirements.txt
cryptography>=41.0.0
pyOpenSSL>=23.0.0
asn1crypto>=1.5.1
pefile>=2023.2.7
python-pkcs11>=0.7.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
pytest>=7.4.0
supabase>=2.0.0
postgrest>=0.13.0
python-dotenv>=1.0.0
serverless-wsgi>=3.0.0

